+++ src/components/organisms/LocationSearchTabs/PortSelectionTab.tsx
@@ -1,8 +1,9 @@
+// src/components/organisms/LocationSearchTabs/PortSelectionTab.tsx
 'use client';
 
 import { Anchor, Loader2 } from 'lucide-react';
 import { useRouter } from 'next/navigation';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useMemo, useState } from 'react';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import {
   Select,
@@ -11,22 +12,9 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@/components/ui/select';
+import type { Port } from '@/hooks/usePortsData';
+import { usePortsData } from '@/hooks/usePortsData';
 import { logger } from '@/lib/logger';
-
-interface Port {
-  id: string;
-  name: string;
-  prefecture_code: string;
-  prefecture_name: string | null;
-  port_code: string;
-  latitude: number | null;
-  longitude: number | null;
-}
-
-interface Prefecture {
-  code: string;
-  name: string;
-}
 
 interface PortSelectionTabProps {
   onPortSelect?: (port: Port) => void;
@@ -34,55 +22,10 @@ interface PortSelectionTabProps {
 
 export function PortSelectionTab({ onPortSelect }: PortSelectionTabProps) {
   const router = useRouter();
+  const { allPorts, prefectures, isLoading, error } = usePortsData();
   const [selectedPrefecture, setSelectedPrefecture] = useState<string>('');
   const [selectedPortId, setSelectedPortId] = useState<string>('');
-  const [allPorts, setAllPorts] = useState<Port[]>([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
 
-  // 初回マウント時に全港を取得
-  useEffect(() => {
-    const fetchAllPorts = async () => {
-      setIsLoading(true);
-      setError(null);
-
-      try {
-        const response = await fetch('/api/ports');
-
-        if (!response.ok) {
-          throw new Error('港情報の取得に失敗しました');
-        }
-
-        const data = await response.json();
-        setAllPorts(data.ports || []);
-      } catch (err) {
-        logger.error({ err }, 'Failed to fetch all ports');
-        setError(err instanceof Error ? err.message : 'エラーが発生しました');
-        setAllPorts([]);
-      } finally {
-        setIsLoading(false);
-      }
-    };
-
-    fetchAllPorts();
-  }, []);
-
-  // 都道府県一覧を生成（港データから重複なしで抽出）
-  const prefectures = useMemo<Prefecture[]>(() => {
-    const prefMap = new Map<string, string>();
-
-    allPorts.forEach((port) => {
-      if (!prefMap.has(port.prefecture_code) && port.prefecture_name) {
-        prefMap.set(port.prefecture_code, port.prefecture_name);
-      }
-    });
-
-    return Array.from(prefMap.entries())
-      .map(([code, name]) => ({ code, name }))
-      .sort((a, b) => parseInt(a.code, 10) - parseInt(b.code, 10));
-  }, [allPorts]);
-
-  // 選択された都道府県の港一覧をフィルタリング
   const filteredPorts = useMemo(() => {
     if (!selectedPrefecture) {
       return [];
@@ -104,7 +47,6 @@ export function PortSelectionTab({ onPortSelect }: PortSelectionTabProps) {
         return;
       }
 
-      // portId で直接 dashboard に遷移
       router.push(`/dashboard?portId=${port.id}`);
 
       if (onPortSelect) {
@@ -116,7 +58,6 @@ export function PortSelectionTab({ onPortSelect }: PortSelectionTabProps) {
 
   return (
     <div className="space-y-3">
-      {/* 都道府県選択 */}
       <Select value={selectedPrefecture} onValueChange={setSelectedPrefecture}>
         <SelectTrigger disabled={isLoading || allPorts.length === 0}>
           <SelectValue placeholder="都道府県を選択" />
@@ -132,19 +73,16 @@ export function PortSelectionTab({ onPortSelect }: PortSelectionTabProps) {
         </SelectContent>
       </Select>
 
-      {/* ローディング */}
       {isLoading && (
         <div className="flex items-center justify-center py-8">
           <Loader2 className="h-6 w-6 animate-spin text-muted-foreground" />
         </div>
       )}
 
-      {/* エラーメッセージ */}
       {error && (
         <div className="rounded-lg bg-destructive/10 p-3 text-sm text-destructive">{error}</div>
       )}
 
-      {/* 港選択 */}
       {!isLoading && filteredPorts.length > 0 && (
         <Select value={selectedPortId} onValueChange={handlePortChange}>
           <SelectTrigger>
@@ -175,7 +113,6 @@ export function PortSelectionTab({ onPortSelect }: PortSelectionTabProps) {
         </Select>
       )}
 
-      {/* 空状態 */}
       {!isLoading && selectedPrefecture && filteredPorts.length === 0 && !error && (
         <div className="text-center py-8 text-sm text-muted-foreground">
           この都道府県に登録されている港がありません
